Updated to v2.1.0
Added in a function that checks is the NPC is already helping a player. If the NPC is already busy the next person will be unable to interact with the bank and simply gets a message to wait for a moment. Making it feel more realistic and they people have to wait for each other like real life.

Also add a 2nd distance check in for the counters. So first it checks if you are in the area ( radius 5 meters ), if in the area it checks your distance from the counter where the NPC is standing.
When close enough the prompt shows up.

This is also so later we people can build on the plugin so you have more NPCs behind the counters if the bank has multiple counters.

NOTE: DokusCore itself needs to update also so you have the config file complete, this will be done sometime during this day. Till then this plugin will show you some errors.